+++ stacktester/nova.py
@@ -1,5 +1,6 @@
 
 import common.http
+import subprocess
 
 class API(common.http.Client):
     def __init__(self, host, port, base_url, user, apikey):
@@ -23,3 +24,20 @@ class API(common.http.Client):
         headers['X-Auth-Token'] = self._authenticate(self.user, self.apikey)
         kwargs['headers'] = headers
         return super(API, self).request(method, url, **kwargs)
+
+    def _add_flavor(self, flavor):
+        #TODO: Get connection info from config
+        subprocess.call(["ssh", "root@nova1", "nova-manage instance_type create", 
+                        str(flavor["name"]),
+                        str(flavor["ram"]),
+                        str(flavor["vcpus"]),
+                        str(flavor["disk"]),
+                        str(flavor["flavorid"]),
+        ]) 
+
+    def _delete_flavor(self, flavor_name):
+        #TODO: Get connection info from config
+        subprocess.call(["ssh", "root@nova1", "nova-manage instance_type delete ",
+                    flavor_name,
+                    "--purge",
+            ])
